refactor: move the wallet package from eudi/wallet to top-level wallet
The standalone EUDI SD-JWT VC wallet is a facade over the eudi/* protocol
libraries, not part of them, so it belongs alongside client rather than under
eudi/. Move eudi/wallet -> wallet (package name unchanged) and update the
walletcli imports, the holderkeys/aliases doc comments, the design docs, and the
changelog. Only the import path changes:
github.com/privacybydesign/irmago/eudi/wallet -> github.com/privacybydesign/irmago/wallet.
