+++ wallet/errors.go
