+++ wallet/policy_test.go
