+++ wallet/wallet_test.go
